[Debug] Removing video urls from abc's fetched articles.
ABC tends to have lots of video "articles" on its website. These throw errors when trying to parse their date, body, etc. As such, they are not included in the POST to the backend. Removing URL's with `/video/` avoids the unnecessary overhead of pulling these video "articles" only to error out.